pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "hyperliquid-mcp-updated"
version = "0.2.1"
description = "MCP server for Hyperliquid Info API, aligned with hyperliquid-python-sdk"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "lucaluca-04" }]
dependencies = [
  "mcp>=1.2.0",
  "hyperliquid-python-sdk>=0.22.0",
]

[project.scripts]
hyperliquid-mcp-updated = "hyperliquid_mcp_updated.server:main"

[tool.hatch.build.targets.wheel]
packages = ["hyperliquid_mcp_updated"]

[tool.hatch.build.targets.sdist]
include = ["hyperliquid_mcp_updated", "README.md", "LICENSE"]
